Fixed a few bugs and small changes
Changed output on admin portal to show building name, and a substring of problem. Fixed default value of ticket submission for updates, along with required users to input their building name, location, and issue. Changed db to show nothing instead of N/A for room number when it's not applicable. Fixed admin portal to be aligned at the top, not center, and added border around each ticket. Shortened down long building names to their abbreviations.